Исправление ревьювера и оформительские изменения кода тестов

+++ main-service/src/main/java/ru/practicum/explorewithme/main/service/CategoryServiceImpl.java
@@ -7,10 +7,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.explorewithme.main.dto.CategoryDto;
 import ru.practicum.explorewithme.main.dto.NewCategoryDto;
+import ru.practicum.explorewithme.main.error.EntityAlreadyExistsException;
+import ru.practicum.explorewithme.main.error.EntityDeletedException;
 import ru.practicum.explorewithme.main.mapper.CategoryMapper;
 import ru.practicum.explorewithme.main.model.Category;
 import ru.practicum.explorewithme.main.error.EntityNotFoundException;
 import ru.practicum.explorewithme.main.repository.CategoryRepository;
+import ru.practicum.explorewithme.main.repository.EventRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -20,12 +23,18 @@ import java.util.stream.Collectors;
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
     private final CategoryMapper categoryMapper;
 
     @Override
     @Transactional
     public CategoryDto createCategory(NewCategoryDto newCategoryDto) {
-        return categoryMapper.toDto(categoryRepository.save(categoryMapper.toCategory(newCategoryDto)));
+        if (!categoryRepository.existsByNameIgnoreCaseAndTrim(newCategoryDto.getName())) {
+            return categoryMapper.toDto(categoryRepository
+                    .save(categoryMapper.toCategory(newCategoryDto)));
+        } else {
+            throw new EntityAlreadyExistsException("Category", "name", newCategoryDto.getName());
+        }
     }
 
     @Override
@@ -34,7 +43,11 @@ public class CategoryServiceImpl implements CategoryService {
         Category category = categoryRepository.findById(categoryId)
                 .orElseThrow(() -> new EntityNotFoundException("Category", "Id", categoryId));
 
-        // Обновляем только имя категории
+        if (categoryRepository.existsByNameIgnoreCaseAndTrim(newCategoryDto.getName()) &&
+                !category.getName().equalsIgnoreCase(newCategoryDto.getName())) {
+            throw new EntityAlreadyExistsException("Category", "name", newCategoryDto.getName());
+        }
+
         if (newCategoryDto.getName() != null && !newCategoryDto.getName().isBlank()) {
             category.setName(newCategoryDto.getName());
         }
@@ -49,8 +62,12 @@ public class CategoryServiceImpl implements CategoryService {
         if (!categoryRepository.findById(categoryId).isPresent()) {
             throw new EntityNotFoundException("Category", "Id", categoryId);
         }
+        if (eventRepository.existsByCategoryId(categoryId)) {
+            throw new EntityDeletedException("Category", "name", categoryId);
+        } else {
+            categoryRepository.deleteById(categoryId);
+        }
 
-        categoryRepository.deleteById(categoryId);
     }
 
     @Override
@@ -59,6 +76,7 @@ public class CategoryServiceImpl implements CategoryService {
         Pageable pageable = PageRequest.of(from / size, size);
         return categoryRepository.findAll(pageable).stream()
                 .map(categoryMapper::toDto)
+                .sorted((c1, c2) -> c1.getName().compareToIgnoreCase(c2.getName()))
                 .collect(Collectors.toList());
     }
